pyproject.toml
[tool.poetry]
name = "common"
version = "0.9.0"
description = "Common utilities for messaging using Azure Service Bus"
authors = ["Devlin Duldulao <webmasterdevlin@gmail.com>"]
readme = "README.md"

packages = [
    { include = "utils" },
    { include = "config" }
]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
azure-servicebus = "^7.12.2"
pika = "^1.3.2"
pytest = "^8.3.2"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

